+++ tsup.config.ts
@@ -12,7 +12,7 @@ export default defineConfig({
   dts: true,
   clean: true,
   splitting: false,
-  sourcemap: true,
+  sourcemap: false,
   external: [
     // Solana
     '@solana/web3.js',